Tela de enceramento adicionada kKKKKKKKKKKKKKK

+++ Hospedaria/fdrQuartos/frmAlteraReserva.cs
@@ -15,7 +15,7 @@ namespace Hospedaria.fdrQuartos
     {
         private ConnectionClass db = new ConnectionClass();
         public Form getform { get; set; }
-        
+
         List<string> listNames = new List<string>();// LISTAS COM OS NOMES
         List<int> idNames = new List<int>();// LISTAS COM OS NOMES
         List<string> listHospedagem = new List<string>();// LISTAS COM OS NOMES
@@ -23,11 +23,11 @@ namespace Hospedaria.fdrQuartos
         List<int> idReserva = new List<int>();// LISTAS COM OS NOMES
         List<DateTime> DataCheckin = new List<DateTime>();// LISTAS COM OS NOMES
         List<DateTime> DataCheckout = new List<DateTime>();// LISTAS COM OS NOMES
-        
+
         public frmAlteraReserva(string _pNomeSuite = "", string _pNomeCliente = "", string _pDatareserva = "", string _pDataSaida = "", char reserva = 'c', bool exit = false)
         {
             InitializeComponent();
-            
+
 
         }
         private void popComboBoxes()
@@ -35,8 +35,8 @@ namespace Hospedaria.fdrQuartos
 
             db.SqlConnection();
             string query = "select hospedagem.nome  as NomeHospedagem, hospedagem.idhospedagem, clientes.nome as NomeCliente , clientes.idclientes,reservas.datareserva, reservas.datasaida, reservas.idreserva from hospedagem inner join reservas on reservas.idhospedagem = hospedagem.idhospedagem inner join clientes on clientes.idclientes = reservas.idclientes  order by reservas.idreserva";
-            db.SqlQuery(query);  Clipboard.SetText(query);
-            
+            db.SqlQuery(query); Clipboard.SetText(query);
+
             SqlDataReader _dr = db.QueryReader();
             if (_dr.HasRows)
             {
@@ -60,13 +60,32 @@ namespace Hospedaria.fdrQuartos
                     cbNomeCheckIn.Items.Add(item);
                 }
                 db.closeConnection();
-                
-    
+                button3.Visible = true;
+                button2.Visible = true;
+                if (idReserva.Count > 0)
+                {
+                    
+
+                }
+                else
+                {
+                    
+                }
+
+
                 cbNomeCheckIn.SelectedIndex = 0;//SELECIONA O PRIMEIRO INDEX PARA SER MOSTRADO LOGO NO INICIO
                 cbQuarto.SelectedIndex = 0;
 
             }
-            
+            else
+            {
+                button3.Visible = false;
+                button2.Visible = false;
+                MessageBox.Show("Nao existem reservas.");
+                this.Close();
+
+            }
+
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -84,23 +103,33 @@ namespace Hospedaria.fdrQuartos
 
         private void button2_Click(object sender, EventArgs e)
         {
-            string query = "delete from reservas where idreserva = '"+idReserva[cbNomeCheckIn.SelectedIndex]+"'";
-            db.SqlConnection();
-            db.SqlQuery(query);  Clipboard.SetText(query);
-            db.QueryRun();
-            db.closeConnection();
-            MessageBox.Show("Reserva cancelada.");
-            this.Close();
+            if (idReserva.Count > 0)
+            {
+
+                string query = "delete from reservas where idreserva = '" + idReserva[cbNomeCheckIn.SelectedIndex] + "'";
+                db.SqlConnection();
+                db.SqlQuery(query); Clipboard.SetText(query);
+                db.QueryRun();
+                db.closeConnection();
+                MessageBox.Show("Reserva cancelada.");
+                this.Close();
+                getform.Visible = true;
+            }
+            else
+            {
+
+                MessageBox.Show("Reserva cancelada.");
+            }
 
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
 
-            string query = "update reservas set idhospedagem = '"+(cbQuarto.SelectedIndex + 1)+ "', idclientes = '"+idNames[cbNomeCheckIn.SelectedIndex]+ "',datareserva = '" + datepicker1.Value.ToString("dd/MM/yyyy HH:mm") + "',datasaida = '" + datepicker2.Value.ToString("dd/MM/yyyy HH:mm") + "'  ";
-            
+            string query = "update reservas set idhospedagem = '" + (cbQuarto.SelectedIndex + 1) + "', idclientes = '" + idNames[cbNomeCheckIn.SelectedIndex] + "',datareserva = '" + datepicker1.Value.ToString("dd/MM/yyyy HH:mm") + "',datasaida = '" + datepicker2.Value.ToString("dd/MM/yyyy HH:mm") + "'  ";
+
             db.SqlConnection();
-            db.SqlQuery(query);  Clipboard.SetText(query);
+            db.SqlQuery(query); Clipboard.SetText(query);
             db.QueryRun();
             db.closeConnection();
             MessageBox.Show("Reserva alterada.");
@@ -108,15 +137,15 @@ namespace Hospedaria.fdrQuartos
 
         private void cbNomeCheckIn_SelectedIndexChanged(object sender, EventArgs e)
         {
-           
+
         }
 
         private void frmAlteraReserva_FormClosing(object sender, FormClosingEventArgs e)
         {
-            //Form1 objPrm1 = new Form1();
-            //objPrm1.Show();
-                
-                
+
+            getform.Visible = true;
+
+
 
         }
     }
